requirements.txt
beautifulsoup4==4.12.3
fastapi==0.109.2
psycopg2_binary==2.9.9
pydantic==2.6.1
python-dotenv==1.0.1
Requests==2.31.0
uvicorn==0.27.1
httpx==0.27.0
langchain==0.1.10
langchain-openai==0.0.8
asyncio==3.4.3
python-telegram-bot==20.8
python-telegram-bot[job-queue]
